\name{Rcpp Modules Examples}
\alias{DataType}
\alias{Rcpp DataType Class}

\title{
  Functions and Objects created by Rcpp Modules Example
}
\description{
  These function and objects are accessible from R via the Rcpp Modules mechanism
  which creates them based on the declaration in the C++ file.  
}
\seealso{
  The Rcpp Modules vignette.
}